requirements.txt
# Web Framework
streamlit==1.32.0

# LLM & Agent Framework (Langchain)
langchain==0.1.13  
langchain-core==0.1.52 
langchain-huggingface==0.0.1
langchain-community==0.0.29

# Vector Database & Embeddings
faiss-cpu==1.8.0
sentence-transformers>=2.6.0

# PDF Processing & Requests
PyMuPDF==1.24.0
requests==2.31.0

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1